:memo: More Arch + Philosophy
Removed the "info" section from index.md because it didn't really make
sense after the talk with Klaus Igleberger. It actually makes more sense
to add back the warning badges but for "ALLOCATES" & "THROWS" to warn
embedded projects that a project could throw an exception or allocate
memory from heap.